Move ToPlannedNodeConverterRegistry to `execution.plan` package
